Updating Assignment Q2
Completed g and h

+++ assignment2/q2_initialization.py
@@ -26,7 +26,7 @@ def xavier_weight_init():
         """
         ### YOUR CODE HERE
         epsilon = tf.sqrt(6.0/ sum(shape))
-        out = tf.Variable(tf.random_uniform(shape, minval=-epsilon, maxval=epsilon, dtype = tf.float32), **kwargs)
+        out = tf.Variable(tf.random_uniform(shape, minval=-epsilon, maxval=epsilon), **kwargs)
         ### END YOUR CODE
         return out
     # Returns defined initializer function.
